% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_similarity.R
\name{calculate_similarity}
\alias{calculate_similarity}
\title{Calculating similarity from HLO-classification}
\usage{
calculate_similarity(classified_data, similarity_weights)
}
\arguments{
\item{classified_data}{A data.frame as returned by function classify_data
with observations in the rows, interactions in the columns and filled
with integers, describing the interaction with respect to the choosen
model assumption.}

\item{similarity_weights}{A set of numbers, how to evaluate the similarity
between different interaction categories ... of class similarity_weights,
e.g. comming from the functions similarity_weights() or MDRDist_weights().}
}
\value{
a similartiy matrix in the form
   [pairwise sum of numerator similarity weights] / [pairwise sum of denominator similarity weights]
   for all pairs of observations over all significant interactions
}
\description{
This function mapps the HLO-classifications to a similarity matrix
}

