% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_mbmdr.R
\name{call_mbmdr}
\alias{call_mbmdr}
\title{Wrapper for MBMDR-4.4.1  (up to the extern function call)}
\usage{
call_mbmdr(formula, data, working_dir)
}
\arguments{
\item{formula}{selecting rows from data}

\item{data}{a data.frame with observations (in the rows), with the columns
containing categories, that are encoded as positive integers. Missing
values are encoded as -9 !}

\item{working_dir}{some empty directory to communicate with MBMDR-4.4.1 and
to store results.}
}
\value{
the path to a (random looking) subdirectory of working_dir,
   in which files were stored
}
\description{
The c++-software MBMDR-4.4.1 expects a specific data format,
   which is read in via harddrive.
   This function stores the data and starts the programme from shell.
}

