% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_weights_def.R
\name{create_weighting_list}
\alias{create_weighting_list}
\title{Describing weights for similarity relation between different interaction categories}
\usage{
create_weighting_list(H_same = 1, H_H = 0, H_O = 0, H_L = 0, O_H = 0,
  O_O = 0, O_L = 0, L_H = 0, L_O = 0, L_L = 0, L_same = 0)
}
\arguments{
\item{H_same}{weight, if both integers are equal and positive}

\item{H_H}{weight, if both integers are positive, but different}

\item{H_O}{weight, if if one integer is positive and one is 0}

\item{H_L}{weight, if if one integer is positive and one is negative}

\item{O_H}{weight, if if one integer is positive and one is 0}

\item{O_O}{weight, if if both integers are 0}

\item{O_L}{weight, if if one integer is negative and one is zero}

\item{L_H}{weight, if if one integer is positive and one is negative}

\item{L_O}{weight, if if one integer is negative and one is zero}

\item{L_L}{weight, if both integers are negative, but different}

\item{L_same}{weight, if both integers are equal and negative}
}
\value{
a list with the input parameters and class label "HLO_weighting",
   which can be used by function similarity_weights
}
\description{
Function that collects weights for all different interaction
   scemes in a HLO-table. Each HLO-cell gets a "model assumption", which
   mapps the letter to an integer. Here weights for any combination of theese
   integers are assigned.
}
\details{
For reasons of symmetry, the parameters should hold
   H_O = O_H, H_L = L_H and L_O = O_L
}
\examples{
# see function MDRDist_weights in file similarity_weights_common.R

}

