% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_assumptions.R
\name{L_same_H_different}
\alias{L_same_H_different}
\title{Model Assumptions: L_same_H_different}
\usage{
L_same_H_different(model)
}
\arguments{
\item{model}{A HLO-Matrix for one interaction}
}
\value{
An integer Matrix of the same size as input with categories, that
   replace the "H", "L" or "O"-Letters.
}
\description{
A model mapping, where each H ends up in a different category and all L in the same
}
\details{
Models are a mapping from a HLO-Matrix to some integer
   categories, where negative values are associated with L, positive values
   are associated with H and 0 (zero) = O (the capital letter).

   There are diffenent "model assumptions" (or model mappings) possible. Each
   uses one single input parameter (a HLO-Matrix) and returns an integer
   matrix of the same size. The desired model assumption is selected by
   function name.
}

