% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdr_dist.R
\name{mdr_dist}
\alias{mdr_dist}
\title{Main function collecting the functions of MDRDist}
\usage{
mdr_dist(data, working_dir, n_rounds, use_existing_models = FALSE,
  evaluate_models = TRUE,
  similarity_weights = MDRDist_weights("first_test"),
  comparison_distribution = compdist_estimated_uncorellated,
  model_assumption = all_different, further_options = NULL)
}
\arguments{
\item{data}{A data.frame with integer categories, that contains the
observations which should be clustered. Missing data are to be encoded
as integer -9, all other categories have to be positive.
If the data contain a column "status", this column will be discarded.}

\item{working_dir}{An existing, empty directory in which to drop intermediate
stages}

\item{n_rounds}{The number of repeated MBMDR-calls}

\item{use_existing_models}{check, whether the working directory is empty and
throw an error, if not – unless former models are allowed by this flag.}

\item{evaluate_models}{an option that stops computation after calling MB-MDR.
This function is intended to be used to calculate models once and run
several with several similarity weights afterwards.}

\item{similarity_weights}{how should different interaction-class-labels be
combined? This is a list of lists as generated by function similarity_weights}

\item{comparison_distribution}{which distribution should be assumed for
generated data? This argument should be passed as function out of the set
( "compdist_estimated_uncorellated", "compdist_uniform")}

\item{model_assumption}{how should different interaction classes be
interpreted with respect to similarity? This argument should be passed as
function out of the set ( "all_same", "all_different")}

\item{further_options}{a structure to manipulate further options.
It is passed as a list of two-element-lists, where the first element describes
one of the parameters in function "set_options()" and the second element
the new value.}
}
\value{
The function returns
   dist: a distance matrix (in dist-structure) and
   importance: a table containing the names and frequency of most appearing
      interactions
}
\description{
Main function collecting the functions of MDRDist
}
\examples{

}

