% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_weights_common.R
\name{MDRDist_weights}
\alias{MDRDist_weights}
\title{Collection of common similarity_weights}
\usage{
MDRDist_weights(weight_name)
}
\arguments{
\item{weight_name}{A character out of the list ("first_test", <to be extended>...)
that references the desired similarity mapping.}
}
\value{
returns an elements of class similarity_weights, that has been
   referenced by input parameter weight_name and can be passed to
   "calculate_similarity".
}
\description{
This function contains common similarity weights, that can be
   used as similarity_weights-parameter in function calculate_similarity.
}
\examples{
my_weight <- MDRDist_weights("first_test")

}

