% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mbmdr.R
\name{read_mbmdr}
\alias{read_mbmdr}
\title{Reading MBMDR-files}
\usage{
read_mbmdr(where_to_find)
}
\arguments{
\item{where_to_find}{The directory in which the MBMDR-files are stored, encoded as \code{character}.
The directories must contain the files named in details below. If \code{where_to_find} is a vector,
each element will be evaluated and the results will be merged.}
}
\value{
A data.frame of class "mbmdr_output", with the columns
\item{\code{$First_Marker}}{ The name of first variable in interaction}
\item{\code{$Second_Marker}}{ The name of second variable in interaction}
\item{\code{$Chi.square}}{ The strength of interaction measured by chi square test}
\item{\code{$p.value}}{ The strength of interaction measured by significance level in maxT-test}
\item{\code{$source}}{ The directory, in which MBMDR stores the files}
\item{\code{$models}}{ A matrix showing the mbmdr-tables.}
}
\description{
The function \code{call_mbmdr} has created some files (including the calculated MBMDR-models) and stored
them on hard drive permanently.

The aim of this function is to reload all data properly, even after
memory has been cleared (e.g. by shutdown of your machine).
}
\details{
In the directory given by parameter \code{where_to_find}, the following files are expected
\itemize{
\item{\code{input.mbmdr} was the source file for MBMDR-4.4.1}
\item{\code{output.table} is a summary of the calculations from MBMDR-4.4.1}
\item{\code{output.models} is a file that encodes the MBMDR-models as linear form}
\item{\code{generating_sample} is a list containing the rownames of the observations used for calculating mbmdr-model}
\item{\code{included.categories} is a set of lists, that contain variable names and the categories used by this variable}
}
Do not edit theese files by hand.
}

