% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{set_options}
\alias{set_options}
\title{Setting parameters used in MDRDist}
\usage{
set_options(command = NA, p = 100, m = 10, x = 0.1, a = "NONE",
  cutting_p = 0.05, fraction_of_real_data = 0.63,
  fraction_of_artificial_data = 0.63, do_bootstrapping = FALSE,
  exclude_creating_data = FALSE, dissimilarity_exponent = 0.5)
}
\arguments{
\item{p}{number of permutations in MBMDR's maxT-test}

\item{x}{significance level in MBMDR-first-level test}

\item{a}{correction-method in MBMDR}

\item{cutting_p}{significance level in MBMDR-second-level test}

\item{fraction_of_real_data}{number of real observations randomly drawn,
where 1 equals nrow(data)}

\item{fraction_of_artificial_data}{number of artificial observations randomly generated,
where 1 equals nrow(data)}

\item{do_bootstrapping}{choosing real observations randomly with
(-> bootstrapping) or without (-> subsampling) replacement}

\item{exclude_creating_data}{should data used for calculating MBMDR be evaluated for distance?}

\item{dissimilarity_exponent}{The exponent used when calculating dissimilarity.
It is "Dissimilarity = (1 - Similarity) ** dissim_exp".
Most commonly it is set to 0.5 or 1.}
}
\value{
nothing, but fills R-intern "options"-list
}
\description{
Setting parameters used in MDRDist
}

