% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_weights_def.R
\name{similarity_weights}
\alias{similarity_weights}
\title{Generating similarity weights for calculating_similarity}
\usage{
similarity_weights(numerator, denominator)
}
\arguments{
\item{numerator}{list of class "HLO_weighting" [which is output of
create_weighting_list], that indicates, how similar classifications should
be rewarded}

\item{denominator}{list of class "HLO_weighting" [which is output of
   create_weighting_list], that indicates, which classification relations
   should be punished

 @details Please note the following
    - For reasons of scaling, it is recommented, that for each element
    numerator < denominator [or at least numerator < denominator * C] for
    some constant.
    - For symmetry, it should not matter, whether the first or the second
    observation is L or H. Therefore you should set H_L = L_H. This applies
    to H_O = O_H and L_O = O_L as well.
    - Further it does not make sense, if a thing is to something more similar
    than to itself. So H_same and L_same should be maximal.}
}
\value{
The function returns a list object with numerator, denuminator and
   class label "similarity_weights".
}
\description{
This function can be used together with create_weighting_list
   to manually assemble similarity weights, which can be used in function
   calculate_similarity. Some usual weightings aur already collected in
   the function MDRDist_weights().
}
\examples{
# see function MDRDist_weights in file similarity_weights_common.R

}

