% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{wait_for_dir}
\alias{wait_for_dir}
\title{Waiting until recently created directory appears}
\usage{
wait_for_dir(Dir, max_wait = 30, timeout = 1)
}
\arguments{
\item{Dir}{path to the directory which we are waiting for}

\item{max_wait}{timeout until assertion will be raised, if dir does not appear}

\item{timeout}{timestep between two attempts to look for the dir}
}
\value{
nothing but certainty, that a directory is callable
}
\description{
Waiting until recently created directory appears
}

